pyproject.toml
[project]
name = "home-dashboard"
version = "0.1.0"
description = "Home dashboard displaying bus arrivals, weather, and air quality"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.2",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]

[project.scripts]
home-dashboard = "home_dashboard.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/home_dashboard"]

[dependency-groups]
dev = [
    "hatchling>=1.28.0",
]
